Adds ability to only SIGTERM the immediate process
This is useful in certain situations (with multiprocessing apps such as
`gunicorn` or `celery`) where we want to signal to the "main" process
only and let *it* handle gracefully terminating the "worker" processes.

In case the "main" process does not gracefully shutdown the "worker"
processes in the given "timeout" we still make sure to send SIGKILL to
all "(-1)" remaining processes.