Removes -s/--single flag
Instead we will signal TERM to the immediate child process by default.
After that we wait for `timeout` period and signal TERM to all children.
Finally, we wait for another `timeout` period and signal KILL to all remaining processes
(in case they didn't exit with the TERM signal by now).

It's very important to set -t/--timeout to the appropriate time based on
the graceful shutdown you allow your processes to linger for.