Update pid1 for latest lts
The original motivation of the PR is to create a new image based on
Ubuntu 22.04 as I want to use 22.04 as base in one of the project.

Some comments on code changes:

- Remove apt update step since it was causing around ~20 MB extra
space. I believe it's better to the upgrade and other steps in the
derived images instead.  - Do some modification to the bash code based
on feedback from shellcheck.